+++ sources/samples/using-client-get.ts
@@ -15,35 +15,23 @@
  */
 
 // tslint:disable:no-console
-import * as http from "../..";
+import { createClient } from "..";
 
-class A {
+const cli = createClient();
 
-    @http.Middleare()
-    public static async writeLogs(
-        ctx: http.RequestContext,
-        next: http.MiddlewareNextCallback
-    ): Promise<void> {
+(async () => {
 
-        console.log(`${ctx.request.method} ${ctx.request.path}`);
+    try {
 
-        return next();
+        const result = await cli.get({
+            url: "https://www.google.com",
+            timeout: 100
+        });
+
+        console.log(result.data.toString());
     }
+    catch (e) {
 
-    @http.Middleare("GET")
-    public static async writeLogs2(
-        ctx: http.RequestContext,
-        next: http.MiddlewareNextCallback
-    ): Promise<void> {
-
-        const resp = ctx.response;
-
-        console.log(`Visited from ${ctx.request.ip}`);
-
-        await next();
-
-        console.log(`Response: ${resp.statusCode} ${resp.statusMessage}`);
+        console.error(e);
     }
-}
-
-export default A;
+})();